+++ CHANGELOG.md
@@ -62,6 +62,7 @@ and this project adheres to
 - glibc: test `__GLIBC_USE()`;
 - shell: add some commands from ostools;
 #### Changed
+- bpf/insn/samples: add `bpf_samples_get_insns_from_string()`;
 - log.mk: rename log functions name;
 - bpf: add `insn/samples/trace_printk.c` from kprobe.c;
 - Move bpf/{bpf.mk,logo.mk} to template/bpf/;